Second version. Calculating num/14 and showing data

+++ DiplomMain/src/main/java/org/example/diplommain/DataModel.java
@@ -1,32 +1,37 @@
 package org.example.diplommain;
 
+import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
+import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
 
 
 public class DataModel {
     private final IntegerProperty id;
-    private final IntegerProperty num;
+    private final DoubleProperty num;
 
-    public DataModel(int id, int num) {
+
+    public DataModel(int id, double num) {
         this.id = new SimpleIntegerProperty(id);
-        this.num = new SimpleIntegerProperty(num);
+        this.num = new SimpleDoubleProperty(num);
     }
 
     public int getId() {
         return id.get();
     }
 
-    public int getNum() {
+    public double getNum() {
         return num.get();
     }
+
 
     public IntegerProperty idProperty() {
         return id;
     }
 
-    public IntegerProperty numProperty() {
+    public DoubleProperty numProperty() {
         return num;
     }
+
 
 }
